requirements.txt
fastapi
uvicorn
scikit-learn
pandas
matplotlib
seaborn
joblib
openai
python-dotenv
notebook
jupyter
faiss-cpu
sentence-transformers
imbalanced-learn

